requirements.txt
dash==2.17.0
dash-daq==0.5.0
pandas==2.2.2
plotly==5.22.0
numpy==1.26.4
paho-mqtt==2.1.0
requests==2.31.0
